Offer to install Benten
This code throws up a dialog box on VS code
in case Benten is not found. This adds to the
complexity of the VS Code extension. It is also
tricky to figure out how to make the installation
commands bullet proof. It is better to rely on
users to manually install Benten as the see fit.

Closes #46